Extract shared JSON value formatter
The Java JACKSON_JSON_NODE and Haskell AESON_VALUE renderers each
defined byte-identical copies of the temporal/bytes/sets/OrderedMap to
``json.dumps``-ready conversion (~30 lines of recursive logic).
Moving the logic to ``_formatters/format_json_value.format_json_value_text``
removes the inconsistent-fix risk Cursor Bugbot flagged and gives any
future ``json_type`` port a single seam.
